+++ src/services/cart/cart-index.js
@@ -1,5 +1,5 @@
 import { fetchCartList } from "./cart-list-api.js";
-import { initCartModalListener } from "./cart-service.js";
+import { initCartModalListener } from "./modal-service.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
     await fetchCartList();
